requirements.txt
Django==3.0.5
djangorestframework==3.11.0
djangorestframework-camel-case==1.1.2
drf-yasg==1.17.1
django-filter==2.2.0
ipython==7.14.0
